[feat] TextFeaturizer and TextNormalizer were moved to base file
