+++ src/core/processes/normalization/normalizers.py
@@ -1,51 +1,17 @@
 from __future__ import annotations
-from abc import abstractclassmethod, abstractmethod
 from typing import List, Union, Callable, Iterable, Generator
 
 import logging
 from functools import reduce
 from omegaconf import OmegaConf
 
-from src.core.interfaces import IProcess
 from src.core.management.managers import CorpusLazyManager
+from src.core.processes.normalization.base import TextNormalizer
 from src.core.processes.normalization.norm_utils import REGEX_NORMALIZATION_HANDLERS
 
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-class TextNormalizer(IProcess):
-    """
-    """
-    def __init__(
-        self, 
-        configs: OmegaConf, 
-        alias: str = None
-        ) -> None:
-        
-        if alias is not None:
-            self._configs = configs.pipeline[alias]
-        else:
-            self._configs = configs
-    
-    @abstractclassmethod
-    def get_isolated_process(cls) -> IProcess:
-        ...
-    
-    @abstractclassmethod 
-    def get_default_configs(cls) -> OmegaConf:
-        """
-        """
-        ...
-    
-    @abstractmethod
-    def normalize_text(
-        self, 
-        corpus: Union[List[str], Iterable]
-    ) -> Union[List[str], CorpusLazyManager]:
-        """
-        """
-        ...
 
 
 class RegexNormalizer(TextNormalizer):
